+++ dubbo-plugin/dubbo-qos/src/main/java/org/apache/dubbo/qos/command/annotation/Cmd.java
@@ -16,13 +16,12 @@
  */
 package org.apache.dubbo.qos.command.annotation;
 
-import org.apache.dubbo.common.utils.StringUtils;
+import org.apache.dubbo.qos.permission.PermissionLevel;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.util.Arrays;
 
 /**
  * Command
@@ -65,50 +64,5 @@ public @interface Cmd {
      * @return command permission level
      */
     PermissionLevel requiredPermissionLevel() default PermissionLevel.PROTECTED;
-
-    enum PermissionLevel {
-        /**
-         * the lowest permission level, can access with
-         * anonymousAccessPermissionLevel=PUBLIC / anonymousAccessPermissionLevel=1 or higher
-         */
-        PUBLIC(1),
-        /**
-         * the middle permission level, default permission for each cmd
-         */
-        PROTECTED(2),
-        /**
-         * the highest permission level, suppose only the localhost can access this command
-         */
-        PRIVATE(3),
-
-        /**
-         * It is the reserved default anonymous permission level, can not access any command
-         */
-        NONE(Integer.MIN_VALUE),
-
-        ;
-        private final int level;
-
-        PermissionLevel(int level) {
-            this.level = level;
-        }
-
-        public int getLevel() {
-            return level;
-        }
-
-        public static PermissionLevel from(String permissionLevel) {
-            if (StringUtils.isNumber(permissionLevel)) {
-                return Arrays.stream(values())
-                    .filter(p -> String.valueOf(p.getLevel()).equals(permissionLevel.trim()))
-                    .findFirst()
-                    .orElse(NONE);
-            }
-            return Arrays.stream(values())
-                .filter(p -> p.name().equalsIgnoreCase(String.valueOf(permissionLevel).trim()))
-                .findFirst()
-                .orElse(NONE);
-        }
-    }
 
 }
